<Refactor>[PDF]: <Better to use data classes>
[Necessitated changes sizes and positions to integer rather than float.  I think this was the dependency upgrade]

[https://github.com/hasii2011/pyumldiagrams/issues/40]

+++ pyumldiagrams/Definitions.py
@@ -63,11 +63,11 @@ class Size:
     """
     Defines the size of the input UML definitions;
     """
-    width:  float = 100
+    width:  int = 100
     """
     The width of a shape
     """
-    height: float = 100
+    height: int = 100
     """
     The height of the shape
     """
